Se modifica el tipo de dato, para los estatus de muestreo. Se cambia de long a int.

+++ Application/Features/Operacion/Muestreos/Commands/Liberacion/DeleteByFilterCommand.cs
@@ -28,7 +28,7 @@ namespace Application.Features.Muestreos.Commands.Liberacion
 
         public async Task<Response<bool>> Handle(DeleteByFilterCommand request, CancellationToken cancellationToken)
         {
-            var data = await _muestreoRepository.GetResumenMuestreosAsync(new List<long> { (long)Enums.EstatusMuestreo.Cargado, (long)Enums.EstatusMuestreo.EvidenciasCargadas });
+            var data = await _muestreoRepository.GetResumenMuestreosAsync(new List<int> { (int)Enums.EstatusMuestreo.Cargado, (int)Enums.EstatusMuestreo.EvidenciasCargadas });
             data = data.AsQueryable();
 
             if (request.Filters.Any())
